+++ doc/CHANGELOG.md
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+0.1.1
+-----
+
+* Added 19 methods to the `Bounds` class,
+3 methods to the `Line` class,
+2 methods to the `Page` class,
+1 method to the `Paragraph` class,
+3 methods to the `TextUtils` class
+
+* Changed `toString` for most classes
+
+* Changed license from LGPL 3 to LGPL 2
+
 0.1
 ---